Move helpers out into own files, to improve maintainability
