+++ spec/spec_helper.rb
@@ -11,28 +11,8 @@ require 'pry'
 require 'webmock/rspec'
 require 'mail'
 require 'sidekiq/testing'
-require 'test_data'
-
-module GotGastro
-  module Env
-    module Test
-      def restore_env
-        ENV.replace(@original) if @original
-        @original = nil
-      end
-
-      def delete_environment_variable(name)
-        @original ||= ENV.to_hash
-        ENV.delete(name)
-      end
-
-      def set_environment_variable(name, value)
-        @original ||= ENV.to_hash
-        ENV[name] = value
-      end
-    end
-  end
-end
+require 'helpers/test_data'
+require 'helpers/env_test'
 
 include GotGastro::Env::Test
